Update e2e test
